Обновление функций, а так же наконец-то проект стал более структурированный
