--- src/navigation/mod.rs
@@ -1,2 +1,0 @@
-pub mod navigation_logic;
-mod navigation_ui;
